Add a test file. At this time, the BES diagnostic has the attenuation of the beam implemented, the reading of the stopping coefficient and the emission coefficient are done, and the loading of all the usefull data from XGC is done.
A small change has been done for reading the ADAS files, now we have an abstract
class that will be herited for all the different kind of ADAS file (only the 
ADF21 and ADF22 will be done for this work, but the file can be easily extend
to other files).

For the collisions class, some names has changed for helping the readers and
the acces methods have now an extra input argument (type of coefficient, for
example if we want an emission coefficient or an attenuation one).

The work is now aimed to compute the real light emission due to the beam,
therefore the beam has to use the data from the class containing the emission
coefficient (should be done quickly).